pdfwrite - when fill->bitmap conversion degenerates to empty, just drop it
Bug #696705 "crash in pdfwrite"

This is caused by a combination of a low resolution and converting
colours into RGB, when a shading uses Separation colour spaces.

The code renders the shading as a bitmap, but the path to fill is empty
triggering the use of the mask instead of the path. However, because
we are a shading not a masked image, there is no mask.

Instead we now simply drop the shading.

No differences expected