BBox device: speed up.
When taking the bbox of a stroked path, don't stroke the entire
path to a new path, and then take the bbox of that. Instead
allow the stroke code to call the underlying bbox code for each
path segment directly.

On files with HUGE paths this can save a lot of mallocs/frees.