Bug 696716: cff parser: cope with incomplete encoded numbers
The fonts in the test case had an incomplete 32 bit number at the end of the
stream - meaning it threw an error trying to read 4 bytes when only 1 byte
was available.

Update the code to read *up to* 4 bytes in these cases.

The fonts are still invalid because they leave a dirty operand stack, but the
cff parser ignores such problems.

Also quieten an uninitialised variable warning.