Add max_used to gs_memory_status and use it for -Z: resource usage
By adding this to the gs_memory_status_t and returning it (if available)
from the gs_memory_status function of allocators, we can print the max
used even when it is not a DEBUG build and avoid the ugly gs_debug
hack previously used to print the max_used for a DEBUG build. The
time for tracking this is microscopic. Also the chunk_mem allocator
now tracks this always, not just for DEBUG builds.