Fix for Bug 696724
This issue was caused by the interaction of the AA alpha and the fact
that we have a knock out group.  In such a case, we don't want the
AA alpha to be knocking out what was already drawn.  Instead we want
to do a blend with the AA alpha and then force the alpha to what ever
the alpha is in the graphic state.  This is achieved by looking at
what is already drawn in our destination and if there is nothing there,
then we just copy.  If there is something there, then we blend and
then set the alpha value appropriately.