txtwrite - Coverity fix, Adobe glyph list lookup not working
Coverity ID 94593

The code to search the Adobe Glyph List in an attempt to find a Unicode
value for a named glyph was incorrect, and simply didn't work. Because
'index' started at -1, and all the searches started by testing if
'index >= 0', all the table searches were skipped.

This should actually run the searches and may improve the text extraction
a little.