Coverity ID 94810 - properly bracket macro
gx_setcurrentpoint is a macro, with 2 lines of code. This meant the else
clause was only applied to the first line, and the second was always
processed.

Fortuitously the if clause would have invalidated the current point so
I believe this is always safe, but its best to be certain and not
corrupt the current point.