pdfwrite - clear/delete extra temp files when run with %d in OutputFile
Bug #696727 "Regression: pdfwrite leaves temp files"

For some reason this does not exhibit on Windows, I have no idea why.
This commit adds a call to the function that closes and removes
pdfwrite's temporary files if we exit due to a '%d' in OutputFile in
order to prevent a spurious extra file being output.

For me this fixes the problem on Linux.