pdfwrite - continue writing fonts even after one fails
File Bug693711.pdf has a number of broken fonts which we cannot write
out (Acrobat throws an error on these pages in the original PDF file).
Previously when we encountered an error writing a font we would stop
writing fonts, even the ones which were valid.

Here we store the error (we don't want to forget there was one, so we
can report it at the end) but carry on emitting the PDF file. This
does a better job of emitting a file from broken input.