Bug 696675 - UI message in output file.
Update this file to be a bit more in sync with its analog in gs
dwmain.c.  Hopefully by redirecting output to stderr we will avoid the
issue reported in the bug where error messages were being directed to
output files.  I was unable to reproduce it locally with Windows 10 once
the changes were applied.