Coverity ID 94589 and 94687 - prevent a NULL dereference
The code contained a poorly named macro (copy_tile) which executed
a device method without testing to see if it was NULL. We know from
prior work that the original intention was that device methods should
*never* be NULL, but we also know that some devices do have NULL methods
and indeed this very routine sets some NULL methods.

So to be safe we now check the method before executing it. At the same
time I've renamed the macro, using upper case so we have some chance of
seeing its a macro, and prevents the macro name collisions with the
code in gdevwddb.c and gdevwprn.c.

I also rewrote it to use if....else instead of ? : because it was
practically unreadable before.

Removed the (pointless) 'return_if_interrupt' macro.

Finally, added the do..while(0) construction around the macro body.