Coverity ID 94578 - avoid dereferencing a pointer if its NULL
The variable 'textures' was checked for being NULL before assigning
traster from it, but we then later used textures without checking it.

I did consider adding a similar check to the original one here, at the
point flagged by Coverity, but the loop would then have proceeded to
use a NULL pointer as an array, leading to further problems.

I suspect that textures cannot (or at least should not) be NULL, and
none of our test files exercise ths case. So I've chosen to simply
return if it ever is.