Coverity ID 94662 - ensure argument is not NULL
gx_make_rop_texture_device checked 'target' to see if it is NULL, but
then later dereferenced target without checking. It 'looks like' this
is a condition which can't happen, because we would get a seg fault.

In fact it is only called from one place, and that location
(gx_image_enum_begin) can return an error, so I've chosen to ensure
that the argument cannot be NULL before calling gx_make_rop_texture_device