Bug #696334 - Don't clean up exec stack after a stackoverflow error
The problem is that the CET file leaves junk on the operand stack after
running the colour transform, it also runs three nested loops, 10 steps
in each loop, for each of the of R, G and B components. This (eventually)
leads to a stack overflow error.

This wouldn't normally be a problem, but we then go round and increase
the stack size, and retry the operation. But, because we had removed the
execution stack operands, the operator failed, leading to a typecheck
error.

We can (and I have) fix this by not popping the elements off the exec
stack when the error is a stackoverflow. This permits us to properly
clean up the exec stack for other types of error, and we can hope that
the stackoverflow is self-correcting. Its not ideal but its the best I
can think of.

No differences expected