Coverity ID 94730 - fix an incorrect index
This one's a puzzle. Coverity is absolutely correct that the '-1'
is invalid, as we use it later as an array index, in fact we possibly
use it in several different sub-functions, as an array index, and its
not valid for it to be negative in any of them. There are no tests
on the value to take special action.

I'm forced to conclude that it should simply be a positive '1' not a
negative '1'.

I am also unable to get the code to take this path, so I can't test the
change, but the -1 was definitely wrong, and a step of 1 looks correct.