Bobbin: Add first version of Bobbin.
A simple tool to help debug performance problems with
threads.

Build with BOBBIN defined, and all pthreads calls (or at
least all the ones gs uses) go through Bobbin. This keeps
track of the time each threads spends waiting on mutexes
or condition variables.

Hopefully this allows contention to be spotted.

A report is printed at the end.