Luratech decoder - set the initial colour space to 'unset'
If the image dictionary has a ColorSpace we set the colorspace member,
and if we get to disconvering the 'N' value, and its still unset then
we set it from the N value. But the Luratech decoder 'set_defaults'
method didn't initially set the colorspace member to 'gs_jpx_cs_unset'
so we might never set it from the /N value.