Bug 696703: Shading disappears at high dpi.
The Shading doesn't disappear, it simply plots solid white.

This is due to an overflow in the tensor patch calculations.

(fixed + fixed + fixed)/3 can overflow.

Change this to be (int64_t + int64_t + int64_t)/3.

For efficiency (and improved accuracy) we postpone the /3 until
the end of the series of operations that use it (where it becomes
a /9).