Build PCL and XPS as shared library and DLL
Also, most of building gpdl as a shared lib and DLL is here, too.

There isn't really an "API" here: the library has one entry point:
"pl_main_aux()", this is really just the build skeleton on which we can
assemble a real API.

On Windows, PCL, XPS and (if forced to build) GPDL will all default to building
a DLL and a "loader" executable (rather than the monolithic exe it build
previously). The monolithic builds can still be done by passing nmake the
MAKEDLL=0 option.

On other (Unix-like) systems, they will default to the monolithic builds.

This all mirrors the gs builds.