Change max_components and num_components in dev.color_info to uchar
There are locations in the code where Coverity has complained about
code logic that resulted in passing of unintialized values that
could only occur if num_components < 0 .  Since this should never
be the case, we will make it clear by setting num_components and
max_components to uchar.   This will limit our support to 256 colors
by a device.