Fix gp_get_realtime commit
The gp_get_realtime commit attempted top use both the upper and lower
values in the long[2] array, this turns out to be a bad idea on Unix
systems, because that uses timeofday() instead of time(), and results
in a time_t which causes a GPF in gmtimte() (NB I think its pretty poor that a
function which takes a 64-bit value can cause a GPF if the value is
sufficiently large!)

Instead use only the bottom long. This will, eventually, cause us to
write incorrect time stamps (when we overflow a long), we will have
to look into that when the time comes.

At the same time fix the definitions of gp_get_realtime() and
gp_get_usertime() in gp.h as these were incorrect.