Bug 696776: Allow gs run without pdfwrite installed.
With the build consolidation change, one of the requirements was that there
should be no dependencies in the graphics library build objects on any of
the interpreter objects.

That meant that gs_pdfwr.ps always gets built in, and then run by the
initialization code, which then throws an error when pdfwrite is not
present.

To handle this, check if pdfwrite is available, and if not, skip executing
the contents of gs_pdfwr.ps.