Font copying - fix the dropping of 'extension' (duplicate) glyphs
Bug #696777 "AddressSanitizer heap-buffer-overflow in copied_drop_extension_glyphs"

When we find that we have a glyph which is processed with different
/Widths, we need to encode it specially in the font. We do this by
extending the glyph name with a ~GS~x where x is an integer.

For output though (pdfwrite and friends) we really don't want to emit
these non-standard glyphs which are duplicates of existing glyphs, the
Widths emission should take care of the metrics problem for us. (in
fact we deal with it by leaving ghe Widths alone and alter the
spacing with a TJ operator).

This is supposed to be performed by 'copied_drop_extension_glyphs()'
but the code for detecting additional duplicates was incorrect and
caused us to read off a block of memory. Fixing that revealed another
problem, the code which was supposed to truncate the 'extended' name
to the normal name was using the wrong loop index and was updating
the 'non-extended' rather than 'extended' glyph name, which of course
had no effect.

I've fixed both cases, and altered the names of the loop counters to
be a little more descriptive in the hope of avoiding future such
problems.

This then exposed faults in the type 1 to Type 2 (CFF) conversion
code which simply wasn't expecting slots to be unused (and as Chris
points out rather begs the question of what this is for). Fixing this
required ratehr more extensive changes to the converter, but it seems
to work correctly now.

No differences expected