Coverity ID 94563: move the fallback assignment.
When a gc run is triggered, if a specific space (local, global etc) isn't
specified, we try to find the space which caused it. In case it wasn't found,
we were setting the pointer to the allocator, but doing so in the wrong place.

Move it to a more suitable place.