Update of ICC creator code
The code had suffered some bit rot due to the movement of MS to
unicode with the newer versions of Visual studio.  I updated the
project to VS 2013, fixed several bugs dealing with unicode, cleaned
the code, and fixed issues in the creation of PS-like CMYK ICC profiles
when using the UCR/BG data.  Updated the UCR/BG data and included an
example with Max K.  Included ICC profiles for Max K and No K.  The
Max K profile would be useful in for text output with -sTextICCProfile=max_k.icc