Coverity ID: 122659 further impossible alpha code tweak
A while back, we identified code in gx_alloc_char_bits() that, due to long
standing changes in the setup code, simply cannot be used. We opted to
conditionally compile out that code, just in case.

Coverity spotted that some later code was dependant on that now redundant code.

So, add the later code to the conditionally compiled out code.