Bug 696800 - subpolygon mode flag not reset.
A flag is used in polygon mode when a subpolygon is created which is
used later when building subpolygon paths (they're special).  The code
that builds the path always resets the flag so the special handling is
not applied to regular paths.  The degenerate case of creating an
empty subpolygon resulted in the flag being left set.  Now we reset
the flag when we exit subpolygon mode.