pdfwrite - honour PDFA-2 restriction on OPM
Bug 696799 " Conformation to PDF/A fails specification ISO 19005-2:2011, clause: 6.2.4.2, test number 2 in veraPDF"

It seems the standards committee decided to disallow OPM 1, essentially.
The spec does say that its only disallowed when a CMYK space is used
and overprinting is true for either stroke or fill, but since that's
the only time its ever used that essentially means it can't be set.

This seems like a bad idea, because PDF/A-1 did *not* enforce this,
which means that the same file cna make a valid PDF/A-1 but not a valid
PDF/A-2.

Still nobody asked my opinion, and its in the spec, so this commit
enforces it. We can't wait until its actually *used*, we have to
prevent it being set in the first place.