txtwrite - fix the ToUnicode processing
commit 9dba57f0f9a53c130ec2771c0ed1d7bd6bbef6ab inadvertently broke
txtweite, I changed the meaning of the return value from the font
decode_glyph method. Initially I had intended it to be a count of
unsigned shorts, but altered it to bytes. Unfortunately I had
changed txtwrite for shorts, but forgot to change it for bytes.

This led to a buffer overrun and heap corruption.