Coverity ID 94614 - remove dead code
lnum can only be 0 on the first time round the loop, initially
num_blank_lines is 0, and is only modified in the 'if' clause, whereas
the test against lnum is in the else clause. So num_blank_lines
cannot be non-zero on the first pass, and lnum cannot be zero on
subsequent passes. Therefore the code is dead.