pdfwrite - prevent invalid combination of encryption and linearisation
The pdfwrite device does linearization as a post-creation process,
this makes it difficult to deal with encryption as we alter the object
numbers, and the object number is used as part of the encryption.

In order to support this we would have to decrypt each string or stream
and then re-encrypt with a different encryption key using the new
object number.

This is way too much effort for a feature of such marginal utility.

Instead, prevent the combination taking place, and document this as a
limitation.