ToUnicode CMaps - avoid buffer overrun
Bug 696811 "Regression: address sanitizer error starting with 9dba57f0f9a53c130ec2771c0ed1d7bd6bbef6ab"

An oversight when I rewrote the ToUnicode CMap processing recently.
Originally the ToUnicode CMap array always carried 2 bytes; now it can
hold an arbitrary length of data.

However, I'd missed the fact that there is one location which expects
to read 2 bytes, if we are only storing 1, and we reach the end of the
array (maximum character code) then we would read one byte past the end
of the array.

Fixed here by checking the size of the value and only reading the second
byte if there are at least 2 (set the second byte value to 0 if not).