Check return code - compiler warning
The code was not checking the return value of sgi_begin_page() which
could fail. If we then continued, various members of 'cur' could
be used uninitialised and in any event the device could not possibly
continue safely.

NB this could happen if we wrote multiple pages and did not specify
'%d' in the OutputFile or if we exhausted memory.