PDF interpreter - refuse to set degenerate text matrix
Bug 696817 "Regression: text missing starting with 83e211723f975beff6ce488a2a6ee5105c089121"

The file sets a font size of 0 and then a ridiculous text matrix:

-2147483648 -2147483648 -2147483648 -2147483648 197 380 Tm

Combined with the CTM this leads to a degenerate CTM (the 2-dimensional
co-ordinates all map to a one dimensional line). This is clearly silly.

This is not in fact a regression, I modified the code so that it would
behave the same when TextAlphaBits is set as when it is not. Although
this file did 'work' previously, when TextAlphaBiots is not set, it
did not work if TextAlphaBits was set, so the code works as expected.

This commit adds a test for a degenerate matrix resulting from Tm and
ignores it if it would. We already 'fix' a font size of 0.