Bug 696824: 'post' table with out of range indices
Loading a Truetype font with a 'post' table which has glyph indices outside
the range of glyph names available, we'd previously exit the loop filling in
the array of names early - leaving subsequent entries in the array unset.
Which would, later, cause a typecheck error.

To be more tolerant, when we encounter an index outside the available range,
just use notdef for that position in the name array.