Slightly hacky pkg-config solution for cross compiling
For cross compiling it is becoming common to include a pkg-config version
specifically tailored for the cross compiling environment, which is
recognized by prefixing the pkg-config tool with a target triplet. For example:
arm-linux-gnueabi-pkg-config

The AC_PATH_TOOL() handles that substitution for us *if* the relevant pkg-config
tool is available.

Unfortunately, if not, it defaults to the "normal" pkg-config tool, which means
we can end up with invalid parameters being setup for the cross compile
environment.

The best solution I could think of was, when cross compiling, we also call the
AC_PATH_PROG() (which ignores cross compiling settings) as well, and compare
the results. If we cross compiling, and both return the same pkg-config
executable, then we don't use pkg-config.