Improve splay tree handling in clumps.
The code walks the splay tree of clumps in 3 main ways.

Firstly it can do an inorder traversal of the tree, from
min to max. We call this a "forward" traversal.

Secondly it can do a reverse-inorder traversal of the tree,
from max to min. We call this a "backward" traversal.

Finally, and most commonly, it can do an inorder traversal
of the tree, from an arbitrary starting position, so that when
it hits the max, the calling code can restart it from the min.

This latter behaviour was nastily implemented, requiring the
callers to jump through hoops. I think that some of the callers
had been getting it wrong, resulting in incomplete traversals
of the tree.

We implement a nicer version here, where the logic for looping
from max to min, and stopping when we reach the start position
is neatly encapsulated in the tree handling.

We also update the clump sanity checking and call it in more
places.

This seems to fix the stale pointer problem that Ray has
encountered, presumably because clumps are correctly freed
now.