ps2write - permit image downsampling
For inline images we set up 'lossless' filters, which use lossless
compression, but also disable downsampling.

This seems wrong for pdfwrite, and because ps2write always uses inline
images, means that downsampling wasn't working at all for ps2write.

This commit alters the lossless filters to allow for image downsampling
even when inline. At the same time we change the ps2write defaults to
not downsample color images, so that we don't see a change in behaviour.