pdfwrite - move linearisation records to non-GC memory
Bug 696832 "crashes on Windows server 2012 R2"

This seems to be because the pointer enumeration was messed up and
we didn't enumerate the 'PageList' at all, causing some spectacular
faults in Garbage collection, because linearization only occurs when
closing the device, which means we always do GC beforehand and move
all the pointers.

The simples solution is just to move the records to non-GC memory, there
is no reason for these to be tracked by the garbage collector.