Memento: Add Memento_tick()
Sometimes it can be useful to be able to run a program
repeatedly and stop it at a given place, just before a
problem occurs.

Suppose you know that the problem occurs in the 'foo'
function, but only after a number of runs.

Insert a call to Memento_tick() at the top of foo, rebuild
and run in the debugger. When the problem occurs, consult
memento.sequence to see what event number we are on;
suppose it's 1000.

Then you can rerun the debugger with breakpoints on Memento_inited
and Memento_breakpoint. When the program stops at Memento_inited,
call Memento_breakAt(1000) and continue execution. The program
will then stop in Memento_breakpoint within the Memento_tick
call just before the problem occurs, enabling you to step forward
and see what goes wrong.