Fix Coverity ID 102148, buffer not terminated.
Also, require the client to pass in a file name without spaces as a
routine precondition instead of trying to replace spaces with
underscores each time a font is built.

Note gs_font_name is a ghostscript string with a size but also requires
null termination, see the typedef's header file.