pdfwrite - don't compress data from Metadata pdfmark when output is PDF/A
Bug 696864 "Metadata stream should not be compressed, when using a pdfmark"

This turned out to be rather more complex than expected. The stream is
constructed via a pdfmark before we add the 'Metadata' key to its
dictionary. This means that at the time we create the stream, we don't
know that it should not be compressed, so we open it with compression
filters added.

We can't throw away the stream we have at the time data is written to
it, because we've written keys to its associated dictionary. Also that
would generate a new object number leaving an unused object in the xref
which is legal but undesirable.

This means that we either have to defer the compression filters
until we write to the file, or close the compression filters when we see
the Metadata. Ideally I'd like to do the former, but I wasn't able to
prove categorically that there was no other code path leading to a
pdfmark-created stream which would evade the checks.

So I've chosen to discard the existing compressed stream when we find
a /Metadata pdfmark, and create a new uncompressed stream for the
stream object. This means also deleting the Filter and DecodeParams
entries from the dictionary, if present.

Finally, there has for some time been an icky hack in the code where we
overload the meaning of 'CompressFonts' to determine whether to compress
all streams outside of page streams (which are controlled with
CompressPages). This commit also adds the new debugging switch
CompressStreams which, if set to false (default is true) will not
compress streams other than Fonts or Pages, which still retain their
existing controls.