Bug 696824 (redux): post table handling
It turns out that it is valid for the number of entries in the post table's name
list to be larger than the number of glyph indices used in the font (for what
purpose, I cannot fathom).

Worse, the number of entries in the name list is not declared anywhere, so the
only option is to pre-process to recover it. In fact, we pro-process the list
of indices (rather than the list of names) and note the highest reference -
this is more efficient as the indices are short, and fixed length values, where
names tend to longer, and of variable length.