pdfwrite - cope with missing optional arguments in /XYZ Dests
Bug 696838 "Preserve /Dests when not named"

Well it seems that 3 of the arguments for an XYZ Destination are
optional and may be omitted, using the current setting instead. The
link destination code wasn't catering for that.

This commit checks the number of arguments in the array, throws errors
when there are too many or not enough, and replaces missing optional
values with 'null' which is equivalent and doesn't break the code.