More TTF post table wranglings
The original TTF spec listed name table indices between 32768 and 65536 as
"reserved for future use". The latest OTF spec (1.60) has indices up to and
including 63335 as valid.

So, tweak our post table handling to cope.

Since numGlyphs is a unsigned short (16 bit) number, there is no point checking
for the upper limit of value, now, so that check is dropped.

Also, since we pre-process the requested indices to ascertain the largest index
we have to deal with, and if the number of names is less than that, fill the
remainder of the array with /.notdef names, we can also drop the explicit check
for indices beyond the available names.