Fix use of shell built-in 'trap'
In instcopy we use the shell built-in 'trap' so that if anything goes wrong
in the copying process, we ensure cleanup afterwards. But we only setup the
trap *after* we'd started copying stuff.

It was *probably* okay, but is safer and cleaner this way.

Noticed in passing looking at Bug 696843.