Match Acrobat's use of AutoRotatePages and DSC
When AutoRotatePages is set to anything except /None Acrobat will
honour DSC Orientation comments and use them in preference to the
heuristically determined value, which we weren't doing.

This commit matches Acrobat, including the ability to have the heuristic
override the DSC comments by setting -dParseDSCComments=false.

The documentation is updated and now correct, including mentioning that
the heuristic can be preferred by turning off DSC parsing.

A few files (7) show a difference because of different orientation
with the pdfwrite device.