PDF interpreter - handle ToUnicode CMaps with large ranges
The CMap reading code did not cope with range entries which spanned
more than 255 indices in a single range, it assumed there would never
be more than this and only used a single byte for the key values.

Fixing that revealed that the ToUnicode generation (from the decoded
CMap) also assumed that a range would never span more than 255 indices.
This was a little more complex to fix.

This commit should allow arbitrary length ranges, even though the
ToUnicode specification currently limits the keys to 0-65536. This is
not tested by cluster runs, but appears to work with all the problematic
ToUnicode files I can find on a quick sweep through Git.