PDF interpreter - Move ExtGState parameters into the regular graphics state
Bug #696511 - "Move ExtGState parameters into the regular graphics state"

A long time ago a decision was made not to replicate the PDF ExtGstate
parameters in the graphics state. This has had long term painful consequences
because we were unable to do a gsave/grestore to simply preserve
parameters but had to jump throuhg a lot of hoops. Amongst other problems
has been confusion over when exactly a change in colour has taken place.
With this commit the ExtGState parameters are moved into the graphics state
and obey gsave and grestore.

This extensive commit was worked on in the PDF_gstate branch and the history
is preserved in that branch.

This does introduce a few differences in the test suite;

There are a few cases where broken files (recursive SMasks) now render
slightly differently.

Some files now exhibit progressions, especially with CMYK or DeviceN output
devices, this is because the 'default' colour space was incorrect, the code
was using DeviceGray instead of the device's colour space, and this makes
a difference for some types of transparency blending when the Group does
not specify a Blending space.

A few other files demonstrate progressions because a change in colour now
takes place before an error occurs.