Initial gstate work, add ca and CA processing as part of the graphics state instead of the //nodict craziness.
Add ca, CA, SMask and AIS to the C graphics state. We'll probably need
more and I think AIS is a duplicate.

With this commit we see a problem in fts25_2526.pdf caused by the lazy
evaluation of SoftMask groups. The output is now incorrect and needs to
be fixed.