PDF interpreter - cope with missing parameters in Destinations
The PDF specification says that certain elements in Destinations can
be 'null' but doesn't say that this actually means 'missing'.

Commit b4fc7327fa0c792a7b218610b86d9fa4533d3e0b added support for this
in XYZ Dests, this commit adds support for FitH, FitV, FitBH and FitBV
Dests completing the variations. Note that the spec says that 'null'
values for FitR have 'undefined results' so we just throw those away
still.

No differences expected.