Add support to create threshold profiles for input color spaces
This adds the ability to the ICC Creator tool in toolbin/color
to create ICC profiles that will threshold to neutral CIELAB
values of black or white.  You can specify the L* point at
which to apply the transition.  The input curve treats the device values as
having the traditional sRGB like gamma.   This is work toward dealing
with the enhancement in Bug 696834