cups-config: rejig detection....
First, allow the CUPSCONFIG variable to be set from the environment - so if
a specific cups-config is available for cross-compiling, you can point
configure to it.

Second, if it's not been set, use AC_PATH_TOOL() to check for it: this will
return a cross-compile compatible cups-config if it's in the search path
and uses the convention of the platform triplet prefix (as we do for
pkg-config).

Lastly, if we're cross compiling, compare what AC_PATH_TOOL() returns with
what AC_PATH_PROG() returns - if they are the same, then no suitable
cross compile cups-config was found (but another was) and we should not
use it (as we do for pkg-config).