CompatibleOverprint blend mode addition (c-code)
This adds support in the transparency code for providing proper
application of blending when we have overprint and transparency.
The description is in Section 7.6.3 of the spec.  This is for
fixing Bug 696876.  This does not complete the problem yet.  We
have to add code in the interpreter side to invoke this blending mode
when overprint is enabled and we are painting elementary graphics objects
(fills, strokes, text, images, and shadings) (That is from the spec).

If the blend mode is anything BUT Normal, the interpreter should
push a non-isolated non-knockout group prior to painting the object. The
blend mode should be set to BLEND_MODE_CompatibleOverprint.  When
we pop the group, the blend mode should be restored to the non-Normal
blend mode for the group composition.