PDF interpreter - cope with Default* ColorSpace definitions in Forms
Bug #696875 " Differences in output - altona swatch p"

Our existing code for dealing with DefaultGray, DefaultRGB and DefaultCMYK
ColorSpace definitions is not great. It relies upon defining these as
ColorSpace resources, and setting UseCIEColor to true. When we then
set DevicGray, DeviceRGB or DeviceCMYK the PostScript code detects the
use of CIEColor and loads the corresponding 'Default' colour space
instead of loading the device space (NB the initial definitions are
identity, ie initally DefaultGray = DeviceGray)

This was only done for Pages, not for Forms, and we now have examples
(test suite files) which use different Default* spaces in Forms.

This code extends the detection of Default* spaces to Form XObjects
and turns on UseCIEColor if it finds any. DoForm now detects any such
spaces and defines appropriate ColorSpace resources. I would like to
put save/restore around the form to preserve any existing definitions,
but that causes the transparency code to seg fault. Instead we carefully
copy any existing definitions, and replace them after running the form.

This show progressions in 2 files:
Altona_Technical_v20_x4.pdf
CATX9509.pdf

The file Bug695948.pdf now times out. This is because of the crazy way
the file is constructed, it has forms nested 50 levels deep, only the
last one making any actual marks. Each form declares every preceding
form as a resource (which it need not do, only the form it actually
uses need be declared). This causes the code attempting to detect the
use of Default* colour spaces to take an extraordinarily long time. I've
decided to accept this as the price for getting the Altona file to work.

In the long term we should rework the use of Default* colour spaces so
that the ICC manager is involved. This should ideally mean storing the
Default* spaces in the graphics state I think.