Support CompatibleOverprint in the interpreter (bug 696876)
If the device is not a HighLevelDevice and the device supports
overprint (ProcessColorModel == /DeviceCMYK) then transparency needs
to treat painting operations differently by pushing a non-isolated,
non-knockout group and set the BlendMode to CompatibleOverprint.
Also the opacityalpha needs to be 1 during painting in this group.

Note that although the interpreter now has code for fill, stroke,
'sh' (shaded fills), text and images, our regression suite does
not seem to have OP with transparency for the 'sh' op (yet).
Out of all of our PDF test files, only 21 trigger this code
(including all of the PDF ATS files that are not run as part of
the commit regression tests.