pdfwrite - fix PDF/X-3 Box emission when boes supplied via pdfmark
Bug #696919 " Wrong TrimBox and Metadata not allowed in PDF/X-3 conversion"

The code to calculate, check and emit the various Boxes when creating
PDF/X output relied upon the mediabox array contents being set. However
this array was initialised to 0, and not set correctly. I suspect this
has been altered at some time in the past, since we now use a temporary
array of floats (instead of doubles) to store and emit the MediaBox.

The upshot of this is that we were writing just the differences from
the MediaBox instead of the full Box values.

No differences expected, the cluster doesn't test PDF/X-3 creation