Fix problems with commit fd34a32 (bug 696841) which transposed clip
Even though the transposition of clip paths commit fd34a32 didn't
turn up any differences, it had problems, some spotted by desk check
and some as a result of testing the customer file and debug.

First, the accum_fill_rectangle needs to transpose coordinates. Then
the clip device must transpose coordinates when comparing to rectangles
in the list. The clip_enumerate function and the procedures that open
code checks against rdev->current transpose the coordinates. If they
need to call clip_enumerate_rest it is with transposed values. The
coordinates in the ccdata structure are non-transposed.

It was surprising when the previous commit didn't show any problems
(my trust in the regression test took a hit).